+++ Tests/KeyboardKitTests/Layout/Providers/StandardKeyboardLayoutProviderTests.swift
@@ -15,7 +15,6 @@ class StandardKeyboardLayoutProviderTests: QuickSpec {
     
     override func spec() {
         
-        var provider: StandardKeyboardLayoutProvider!
         var inputSetProvider: MockInputSetProvider!
         var context: KeyboardContext!
         var device: MockDevice!
@@ -29,22 +28,25 @@ class StandardKeyboardLayoutProviderTests: QuickSpec {
             inputSetProvider.alphabeticInputSetValue = AlphabeticInputSet(rows: [["a", "b", "c"], ["a", "b", "c"], ["a", "b", "c"]].map(InputSetRow.init))
             inputSetProvider.numericInputSetValue = NumericInputSet(rows: [["1", "2", "3"], ["1", "2", "3"], ["1", "2", "3"]].map(InputSetRow.init))
             inputSetProvider.symbolicInputSetValue = SymbolicInputSet(rows: [[",", ".", "-"], [",", ".", "-"], [",", ".", "-"]].map(InputSetRow.init))
-            provider = StandardKeyboardLayoutProvider(
+        }
+
+        func provider(for device: DeviceType) -> StandardKeyboardLayoutProvider {
+            StandardKeyboardLayoutProvider(
+                device: device,
                 inputSetProvider: inputSetProvider,
                 dictationReplacement: .primary(.go))
         }
         
-        
         describe("keyboard layout provider for context") {
             
             it("is phone provider if context device is phone") {
-                device.userInterfaceIdiomValue = .phone
+                let provider = provider(for: .phone)
                 let result = provider.layoutProvider(for: context)
                 expect(result).to(be(provider.iPhoneProvider))
             }
             
             it("is pad provider if context device is pad") {
-                device.userInterfaceIdiomValue = .pad
+                let provider = provider(for: .pad)
                 let result = provider.layoutProvider(for: context)
                 expect(result).to(be(provider.iPadProvider))
             }
@@ -53,7 +55,7 @@ class StandardKeyboardLayoutProviderTests: QuickSpec {
         describe("keyboard layout for context (just testing this one)") {
             
             it("is phone layout if context device is phone") {
-                device.userInterfaceIdiomValue = .phone
+                let provider = provider(for: .phone)
                 let layout = provider.keyboardLayout(for: context)
                 let phoneLayout = provider.iPhoneProvider.keyboardLayout(for: context)
                 let padLayout = provider.iPadProvider.keyboardLayout(for: context)
@@ -62,7 +64,7 @@ class StandardKeyboardLayoutProviderTests: QuickSpec {
             }
             
             it("is pad layout if context device is pad") {
-                device.userInterfaceIdiomValue = .pad
+                let provider = provider(for: .pad)
                 let layout = provider.keyboardLayout(for: context)
                 let phoneLayout = provider.iPhoneProvider.keyboardLayout(for: context)
                 let padLayout = provider.iPadProvider.keyboardLayout(for: context)
@@ -75,6 +77,7 @@ class StandardKeyboardLayoutProviderTests: QuickSpec {
             
             it("changes the provider instance for all providers") {
                 let newProvider = MockInputSetProvider()
+                let provider = provider(for: .phone)
                 provider.register(inputSetProvider: newProvider)
                 expect(provider.inputSetProvider).toNot(be(inputSetProvider))
                 expect(provider.inputSetProvider).to(be(newProvider))